rocksdb: enable O_DIRECT for WAL files
we accumulate stuff until Sync() or until we run out of room in our
currently 512k buffer.

we also stash "max written offset" for each WAL in shared mem and
then on reread, truncate the file back down to that length.  this
should ensure we don't end up with incomplete records on disk during
application crash, but all bets are off for power cut failover.

Signed-off-by: Joel Jacobson <joel.jacobson@primarydata.com>